+++ R/BumpRiverStage.R
@@ -11,7 +11,7 @@
 #' @param min.drop numeric.
 #'   Minimum drop in stage between adjacent river cells.
 #'
-#' @details The \href{http://en.wikipedia.org/wiki/Lee_algorithm}{Lee algorithm} (Lee, 1961)
+#' @details The \href{https://en.wikipedia.org/wiki/Lee_algorithm}{Lee algorithm} (Lee, 1961)
 #'   is used to identify flow paths among the modeled river cells.
 #'   An analysis of river cell stage values along a flow path identifies any problematic cells
 #'   that are obstructing downhill surface-water flow.
@@ -93,7 +93,7 @@ BumpRiverStage <- function(r, outlets, min.drop=1e-06) {
   source.cells <- unlist(source.cells)
 
   # Find paths by finding the shortest path through the river maze.
-  # Lee algorithm, http://en.wikipedia.org/wiki/Lee_algorithm
+  # Lee algorithm, https://en.wikipedia.org/wiki/Lee_algorithm
 
   CalcWaveExpansion <- function(cells) {
     visited.cells <- NULL