+++ R/ExportRasterStack.R
@@ -29,7 +29,6 @@
 #'
 #' @return Used for the side-effect files written to disk.
 #'
-#'
 #' @author J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
 #'
 #' @seealso \code{\link{writeRaster}}
@@ -85,7 +84,7 @@ ExportRasterStack <- function(rs, path, zip="",
   save(rs, file=f)
 
   f <- file.path(path.kml, "rasters.kml")
-  crs <- "+proj=longlat +datum=WGS84"
+  crs <- "+init=epsg:4326"
   rs <- projectRaster(rs, crs=crs, method="ngb", alignOnly=FALSE)
   KML(rs, f, col=col, maxpixels=ncell(rs) * 2, blur=5, zip=zip,
       overwrite=TRUE)